fix(outreach): type group-report visit ordinal explicitly

+++ supabase/migrations/20260826185000_group_report_guest_outreach_contract.sql
@@ -77,7 +77,7 @@ begin
     p_phone,
     p_email,
     p_language,
-    1,
+    1::smallint,
     v_report.meeting_date
   ) x;
